requirements.txt
fastapi
uvicorn
google-genai
pydantic
pydantic-settings
python-dotenv
aiofiles
